requirements.txt
streamlit
pandas
python-docx
pymupdf
google-generativeai
python-dotenv
